backend/requirements.txt
# nnUNet requirements - comment these in when you're ready to integrate the real model
nnunetv2
torch
torchvision
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
nibabel==5.1.0
numpy==1.26.4
Pillow==10.1.0
python-dotenv==1.1.0
celery==5.3.4
django-celery-results==2.5.1
django-filter==23.3
drf-yasg==1.21.7
django-unfold==0.11.0
django-colorfield==0.9.0
scipy==1.11.4
setuptools>=65.0.0
nibabel